+++ R/copernicus_article.R
@@ -13,7 +13,7 @@
 #'
 #' An number of required and optional manuscript sections, e.g. `acknowledgements`, `competinginterests`, or `authorcontribution`, must be declared using the respective properties of the R Markdown header - see skeleton file.
 #'
-#' **Version:** Based on `copernicus_package.zip` in the version 6.4, 8th August 2021, using `copernicus.cls` in version 9.32, 7 July 2021.
+#' **Version:** Based on `copernicus_package.zip` in the version 6.7, 16 March 2022, using `copernicus.cls` in version 9.46, 25 March  2022.
 #'
 #' **Copernicus journal abbreviations:** You can use the function `copernicus_journal_abbreviations()` to get the journal abbreviation for all journals supported by the Copernicus article template.
 #'
@@ -48,21 +48,23 @@
 #' render("MyArticle/MyArticle.Rmd")
 #' }
 #' @export
-copernicus_article <- function(
-  ..., keep_tex = TRUE, highlight = NULL, citation_package = "natbib", md_extensions = c(
-    "-autolink_bare_uris", # disables automatic links, needed for plain email in \correspondence
-    "-auto_identifiers"    # disables \hypertarget commands
-  )
-) {
-  if ("extra_dependencies" %in% names(list(...)))
+copernicus_article <- function(..., keep_tex = TRUE, highlight = NULL,
+                               citation_package = "natbib",
+                               md_extensions = c(
+                                 "-autolink_bare_uris", # disables automatic links, needed for plain email in \correspondence
+                                 "-auto_identifiers" # disables \hypertarget commands
+                               )) {
+  if ("extra_dependencies" %in% names(list(...))) {
     warning(
       "Copernicus does not support additional LaTeX packages and options!
           >> Please remove 'extra_dependencies' from your YAML header!",
       call. = FALSE
     )
+  }
 
   pdf_document_format(
-    "copernicus", citation_package = citation_package,
+    "copernicus",
+    citation_package = citation_package,
     keep_tex = keep_tex, highlight = highlight, md_extensions = md_extensions, ...
   )
 }
@@ -87,6 +89,7 @@ copernicus_journals <- list(
   "Earth System Dynamics" = "esd",
   "Earth System Science Data" = "essd",
   "E&G Quaternary Science Journal" = "egqsj",
+  "EGUsphere" = "egusphere",
   "European Journal of Mineralogy" = "ejm",
   "Fossil Record" = "fr",
   "Geochronology" = "gchron",
@@ -121,6 +124,7 @@ copernicus_journals <- list(
 #' @export
 copernicus_journal_abbreviations <- function(journal_name = "*") {
   unlist(copernicus_journals[grep(
-    journal_name, names(copernicus_journals), ignore.case = TRUE
+    journal_name, names(copernicus_journals),
+    ignore.case = TRUE
   )])
 }